Remove unused Git attributes ident
The $Id$ keywords were in general used in CVS and Subversion where they
can be substituted with filename, last revision number change, last changed
date, and last user who changed it.

In Git this functionality is different and can be done with Git attribute
ident. These need to be defined manually for each file in the
.gitattributes file and are afterwards replaced with 40-character
hexadecimal blob object name which is based only on the particular file
contents.

This patch simplifies handling of $Id$ keywords by removing them since
they are not used anymore and are outdated.