Replace AM_CONFIG_HEADER with AC_CONFIG_HEADERS
The AM_CONFIG_HEADER macro is part of automake and has been removed in
version 1.13.0 and then readded back in the 1.13.2. Originally it was
superseeded by the AC_CONFIG_HEADERS macro since July 2002. Since most
likely this macro will be removed in the future versions it probably
makes sence to now replace it. Also the configure script is pregenerated
and shipped with OpenLiteSpeed so the minimal requirements are even less
problematic here.